fix: cast away MSVC C4244/C4267 narrowing warnings on intentional truncations
All are cases where the wider value is already known to fit: a packed
monochrome glyph row, a fixed-point coordinate seeded from an int
control point, a pixel value already reduced to its target bit depth,
a bitfifo offset bounded by its own (unsigned int) capacity, a
PackBits run length bounded by 128, and a file offset narrowed to the
long that fseek expects. No behaviour changes, just explicit casts.

+++ libraries/datastruct/bitfifo/bitfifo.c
@@ -157,7 +157,7 @@ result_t bitfifo_enqueue(bitfifo_t       *fifo,
            newbitsoffset,
            nnewbits);
 
-    fifo->head += nnewbits;
+    fifo->head += (unsigned int) nnewbits;
   }
   else
   {
